+++ CHANGELOG.md
@@ -47,6 +47,7 @@ The format is based on [Keep a Changelog 1.0.0].
 - users return to the same place in the task list after answering a question
 - fix extended radio questions so that further_information can be remembered when it can be provided through multiple fields
 - flash messages can be shown in notification banners
+- fix checkbox answer to save choices after previously skipping
 
 ## [release-005] - 2021-1-19